+++ CHANGELOG.md
@@ -634,6 +634,11 @@ $$Pow({x}, {n})=Pow\left({x}^2, \dfrac{n}{2}\right) * Pow\left(x, n \mod 2\right
 
 - Base 26
 
+## 0169 - Majority Element
+
+- Map
+- TODO: Boyer Moore
+
 ## 0415 - Add strings
 
 - Full adder